requirements.txt
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.3.2
numpy>=1.21.0
